Unify all Lua scripts into a single file
